requirements.txt
Flask==2.2.5
Flask-WTF==1.0.1
Werkzeug==2.2.3
WTForms==3.0.1
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
requests==2.31.0
email_validator
